+++ src/main/kotlin/app/revanced/patches/youtube/ad/video/patch/VideoAdsPatch.kt
@@ -19,7 +19,7 @@ import app.revanced.patches.youtube.misc.integrations.patch.IntegrationsPatch
 @Patch
 @Dependencies([IntegrationsPatch::class])
 @Name("video-ads")
-@Description("Removes ads in the YouTube video player.")
+@Description("Removes ads in the video player.")
 @VideoAdsCompatibility
 @Version("0.0.1")
 class VideoAdsPatch : BytecodePatch(