+++ src/main/kotlin/app/revanced/patches/youtube/layout/branding/header/patch/PremiumHeadingPatch.kt
@@ -19,7 +19,7 @@ import kotlin.io.path.exists
 @Patch
 @Dependencies([FixLocaleConfigErrorPatch::class])
 @Name("premium-heading")
-@Description("Shows premium branding on the YouTube home screen.")
+@Description("Shows premium branding on the home screen.")
 @PremiumHeadingCompatibility
 @Version("0.0.1")
 class PremiumHeadingPatch : ResourcePatch() {